Add examples for an echo and proxy HTTP server
To showcase how the new APIs are working this PR adds an example for an echo and proxy HTTP server. Right now the examples are not runnable because we are lacking a concrete HTTP server implementation but it shows this working on top of the protocols.

The proxy case is interesting since it shows how nicely the reader and writer APIs fit into each other.