Introduce a `ClosureHTTPClientEventHandler` to allow developers to pass closures
## Motivation

Requiring users to implement a custom `HTTPClientEventHandler` to customize individual event handling is quite cumbersome so we want to provide an easier closure based approach.

## Modifications

This PR adds a new `ClosureHTTPClientEventHandler` and an extension on `HTTPClientEventHandler` that makes it easy for users to pass custom closures instead of creating an event handler themselves.

## Result

We now make it easy for users to customize the event handling without having to duplicate the API surface.